+++ backend/src/api/contact/contact.module.ts
@@ -3,9 +3,10 @@ import { ContactService } from './contact.service';
 import { ContactController } from './contact.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Contact } from './entities/contact.entity';
+import { CloudinaryModule } from '../cloudinary/cloudinary.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Contact])],
+  imports: [TypeOrmModule.forFeature([Contact]), CloudinaryModule],
   controllers: [ContactController],
   providers: [ContactService],
 })
