+++ backend/src/api/contact/dto/create-contact.dto.ts
@@ -1,5 +1,5 @@
-import { IsNotEmpty, IsString } from 'class-validator';
-import { ApiProperty } from '@nestjs/swagger';
+import { IsNotEmpty, IsOptional, IsString } from 'class-validator';
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 export class CreateContactDto {
   @IsString()
   @IsNotEmpty()
@@ -22,7 +22,7 @@ export class CreateContactDto {
   content: string;
 
   @IsString()
-  @IsNotEmpty()
-  @ApiProperty({ description: 'File ID of the contact', example: '123e4567-e89b-12d3-a456-426614174000' })
+  @IsOptional()
+  @ApiPropertyOptional({ description: 'File ID of the contact', example: '123e4567-e89b-12d3-a456-426614174000' })
   fileId: string;
 }
