+++ frontend/app/(private)/manage/report/components/DialogReport.tsx
@@ -1,5 +1,4 @@
-import { IReview } from '@/api/review/interface';
-import StarRating from '@/app/(root)/company/components/StarRating';
+import Link from 'next/link';
 import { Button } from '@/components/ui/button';
 import {
   Dialog,
@@ -11,7 +10,7 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/components/ui/dialog';
-import { Eye } from 'lucide-react';
+import { ExternalLink, Eye } from 'lucide-react';
 import useGetContactById from '../hooks/useGetContactById';
 import { formatDate } from 'date-fns';
 interface IProps {
@@ -20,7 +19,7 @@ interface IProps {
 export default function DialogReport({ id }: IProps) {
   const { data } = useGetContactById(id);
   if (!data) return null;
-  const { fullName, email, title, content, createdAt } = data;
+  const { fullName, email, title, content, createdAt, file } = data;
   return (
     <Dialog>
       <DialogTrigger asChild>
@@ -35,7 +34,7 @@ export default function DialogReport({ id }: IProps) {
             Báo cáo của <span className="font-bold">{fullName}</span>
           </DialogDescription>
         </DialogHeader>
-        <div className="grid sm:grid-cols-2 gap-4">
+        <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
           <div>
             <label className="text-sm font-medium">Tên người báo cáo:</label>
             <p>{fullName}</p>
@@ -55,11 +54,21 @@ export default function DialogReport({ id }: IProps) {
             <label className="text-sm font-medium">Ngày tạo:</label>
             <p>{formatDate(createdAt, 'dd/MM/yyyy')}</p>
           </div>
-          <div className="col-span-2">
+
+          <div className="sm:col-span-2">
             <label className="text-sm font-medium">Nội dung:</label>
             <p className="mt-1 text-sm bg-gray-100 dark:bg-gray-800 p-2 rounded-md border-l-2 border-green">
               {content}
             </p>
+          </div>
+
+          <div className="sm:col-span-2">
+            <label className="text-sm font-medium">File đính kèm:</label>
+
+            <Link href={file.url} target="_blank" className="text-sm text-blue-500 flex items-center gap-2">
+              {file.name}
+              <ExternalLink className="h-3 w-3" />
+            </Link>
           </div>
         </div>
